pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "logging>=0.4.9.6",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "trafilatura>=2.0.0",
]
